Remove digest typehash and reuse message instead
Dapp here is a generic non-Sequence aware dapp, but handles both EOA and ERC-1271 signatures:

- dapp sends `eth_sign('hello world!')` request to wallet
- dapp expects that the digest that was signed was `keccak256('\x19Ethereum Signed Message:\n12hello world!')`
- it sees that the signature is not 65 bytes (i.e. not an EOA signature)
- so it calls `isValidSignature(keccak256('\x19Ethereum Signed Message:\n12hello world!'), signature)`

For Sequence v3, `isValidSignature` validates against a `KIND_DIGEST` payload with whatever digest was passed in.

This means in order for our `eth_sign` to conform, we must sign:
  - a `KIND_DIGEST` payload where digest is `keccak256('\x19Ethereum Signed Message:\n12hello world!')`

In EIP-712, `bytes` and `string` are encoded as the `keccak256` of their contents.

So we could make it work by:
  - changing the type encoding to `Digest(bytes message,address[] wallets)` (or just using `Message`)
  - sign the message `'\x19Ethereum Signed Message:\n12hello world!'` for the `eth_sign` request

For `eth_signTypedData`:

The dapp assumes the hash to validate against is `keccak256('\x19\x01' | domainSeparator | hashStruct(message))`

So we should just be able to sign:
  - a `KIND_DIGEST` payload where the message is `'\x19\x01' | domainSeparator | hashStruct(message)`

Since message already accepts a `bytes`, we can use that instead.